backend/requirements.txt
fastapi==0.115.10
pydantic==2.10.6
pydantic-settings==2.10.1
pydantic[email]
uvicorn==0.34.0
uuid_v7
jinja2
SQLAlchemy[asyncio]==2.0.41
asyncpg==0.29.0
thefuzz
inflect
redis
celery
passlib==1.7.4
bcrypt==4.0.1
python-jose==3.4.0
dotenv
colorlog
httpx
httpx_aws_auth
pillow
python-multipart
